Move pack to its own module
