+++ src/mirage/middleware/patch-auto-type-resolvers.ts
@@ -6,10 +6,11 @@ import {
   isUnionType,
   isInterfaceType,
 } from 'graphql';
-import { PackOptions, ResolverMap } from '../../types';
+import { ResolverMap } from '../../types';
 import { mirageUnionResolver } from '../resolvers/union';
 import { mirageInterfaceResolver } from '../resolvers/interface';
-import { embedPackOptionsInContext } from '../../resolver-map/utils';
+import { PackOptions } from '../../pack/pack';
+import { embedPackOptionsInContext } from '../../pack/utils';
 
 export function patchAutoTypeResolvers(resolverMap: ResolverMap, packOptions: PackOptions): ResolverMap {
   const { graphqlSchema: schema } = packOptions.dependencies;
