+++ src/pack/index.ts
@@ -1,0 +1,1 @@
+export { pack } from './pack';
